+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.1.0] - 2022-04-29
+
 ### Added
 
 - [x] Add new artifacts.
@@ -14,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - [x] Edit versions state.
 - [x] Delete artifacts.
 - [x] Add setting to display name (if exist) instead of ID in registry explorer view.
+- [x] Add setting to format document on preview.
 
 ### Updated
 
@@ -21,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - [x] Fix preview issue.
 - [x] Fix kconnect icon issue.
 - [x] Enhance README with screen records.
+- [x] Manage [Apicurio issue](https://github.com/Apicurio/apicurio-registry/issues/2143) on YAML (OAS) mime type.
 
 ## [v0.0.4] - 2022-02-17
 
@@ -59,7 +63,8 @@ All notable changes to this project will be documented in this file.
 
 - [x] Basic Apicurio explorer.
 
-[unreleased]: <https://github.com/jetmartin/apicurio/compare/v0.0.4...HEAD>
+[unreleased]: <https://github.com/jetmartin/apicurio/compare/v0.1.0...HEAD>
+[c0.1.0]: <https://github.com/jetmartin/apicurio/compare/v0.0.4...v0.1.0>
 [v0.0.4]: <https://github.com/jetmartin/apicurio/compare/v0.0.3...v0.0.4>
 [v0.0.3]: <https://github.com/jetmartin/apicurio/compare/v0.0.2...v0.0.3>
 [v0.0.2]: <https://github.com/jetmartin/apicurio/compare/v0.0.1...v0.0.2>